refactor: module, constant 분리
