+++ constant.js
@@ -1,0 +1,5 @@
+const ERROR = {
+  WRONG_COMMAND: "제대로 된 명령어를 입력해주세요.",
+};
+
+module.exports = { ERROR };
