+++ todos.js
@@ -1,16 +1,3 @@
-let todos = [
-  {
-    name: "자바스크립트 공부하기",
-    tags: ["programming", "javascript"],
-    status: "todo",
-    id: 12123123,
-  },
-  {
-    name: " 그림 그리기",
-    tags: ["picture", "favorite"],
-    status: "doing",
-    id: 312323,
-  },
-];
+let todos = [];
 
 module.exports = { todos };
